Split punctuation by default
Insert a {\k0} in front of any punctuation (excluding apostrophes) to
exclude it from the surrounding syllables. Doing so produces a nicer
effect (see the OP of the 0x539 release of "Kobayashi-san Chi no Maid
Dragon" and the OP of the SkelterSubs release of "Tenkuu Danzai
Skelter+Heaven").

The previous behaviour can be accessed by passing the -j flag.