Deprecate -j in favour of experimental -p
A deprecation notice has been added for -j and will be removed in v1.

Split punctuation is unexpected enough that it should not be turned on
by default. It causes issues with \kf templates and may cause issues in
noblank templates. For \kf templates, the \kf0 syllable is fully
highlighted as if the cursor has already reached there.

In addition, the way the script splits punctuation is currently
broken. The current method of splitting syllables is only robust for
end-of-syl punctuation. Given a syllable `|"The`, the script will
produce output similar to the following: `{\k10}{k0}"The`, whereas the
proper output should be `{\k0}"{\k10}".